+++ tests/chord_tests.py
@@ -14,55 +14,8 @@ from thrift.server import TServer
 
 from contextlib import contextmanager
 from helpers import encode_node
+from TestHelpers import *
 
-
-@contextmanager
-def connect(port):
-    transport = TTransport.TBufferedTransport(
-            TSocket.TSocket('localhost', port))
-    protocol = TBinaryProtocol.TBinaryProtocol(transport)
-    client = KeyValueStore.Client(protocol)
-    transport.open()
-    yield client
-    transport.close()
-
-
-def start_server_with_name_port(port, chord_name=None, chord_port=None):
-    handler = ChordServer('localhost', port, chord_name, chord_port)
-    processor = KeyValueStore.Processor(handler)
-    transport = TSocket.TServerSocket('localhost', port)
-    tfactory = TTransport.TBufferedTransportFactory()
-    pfactory = TBinaryProtocol.TBinaryProtocolFactory()
-    server = TServer.TThreadedServer(processor, transport, tfactory, pfactory)
-    server.serve()
-
-
-def spawn_server(port, chord_name=None, chord_port=None):
-    p = multiprocessing.Process(target=start_server_with_name_port,
-            args=(port, chord_name, chord_port))
-    p.start()
-    sleep(1)
-    return p
-
-
-def create_servers_in_range(from_port, to_port):
-    ports = range(from_port, to_port)
-    servers = {}
-
-    for port in ports:
-        servers[port] = {}
-        if port == ports[0]:
-            servers[port] = spawn_server(port)
-        else:
-            servers[port] = spawn_server(port,
-                                         chord_name="localhost",
-                                         chord_port=ports[0])
-        sleep(2)
-
-    for port in ports:
-        with connect(port) as client:
-            client.print_details()
-    return servers
 
 
 class TestChord:
@@ -227,99 +180,6 @@ class TestChord:
             for port, name in servers.items():
                 name.terminate()
 
-    def test_data_replication(self):
-        #For each server, check that the successor has its data
-        ports = range(3342, 3345)
-        servers = create_servers_in_range(3342, 3345)
-
-        try:
-            #For each port, get its successor
-            for port in ports:
-                print "current port is %d\n" % port
-                key = str(port) + "-key"
-                value = str(port) + "-value"
-                with connect(port) as client:   
-                 
-                    #For each server, put a key that should be sent to its successor
-                    status = client.put(key, value)
-                    assert status == ChordStatus.OK
-
-                    #Now find the node that is the master for the key we just put
-                    node_key = client.get_successor_for_key(str(get_hash(key)))
-                    new_port = int(node_key.split(":")[1])
-
-                    #Check the new_port's successor to see if it stored the replicate
-                    if new_port == port:
-                        #If the key was stored on port (ourselves)
-                        succ = client.get_successor()
-                    else:
-                        #Otherwise, get successor
-                        with connect(new_port) as client:
-                            succ = client.get_successor()
-
-                    #Connect to the successor that should have the replicate
-                    succ_port = int(succ.replace("localhost:", ""))
-                    with connect(succ_port) as client:
-                        response = client.get_replicate_list()
-                        #It should only have one source
-                        assert len(response.replicate_list) == 1
-
-                        #Make sure the right source (port) is in the list
-                        find_result = str(response.replicate_list).find(str(new_port))
-                        print find_result
-                        assert find_result != -1
-
-        finally:
-            for port, name in servers.items():
-                name.terminate()
-
-    def test_crash_some_servers(self):
-        ports = range(3342, 3347)
-        servers = {}
-
-        for port in ports:
-            if port == ports[0]:
-                servers[port] = spawn_server(port)
-            else:
-                servers[port] = spawn_server(port,
-                                             chord_name="localhost",
-                                             chord_port=ports[0])
-
-        for port in ports:
-            with connect(port) as client:
-                client.print_details()
-                
-        try:
-            with connect(ports[0]) as client:
-                client.put('connie', 'lol')
-                client.put('rakesh', 'lol++')
-                key = client.get_successor_for_key(str(get_hash('connie')))
-                
-                print "key is " + key
-                print "hash is " + str(get_hash('connie'))
-                port = int(key.split(":")[1])
-                resp = client.get('connie')
-                assert resp.value == 'lol'
-
-            servers[port].terminate()
-            del servers[port]
-
-            # Can get the data from somewhere (replication).
-            with connect(ports[0]) as client:
-                resp = client.get('connie')
-                assert resp.value == 'lol'
-
-            # New keys still get propagated around.
-            with connect(ports[0]) as client:
-                client.put('russell', 'organic')
-
-            with connect(ports[1]) as client:
-                resp = client.get('russell')
-                assert resp.value == 'organic'
-
-        finally:
-            for port, name in servers.items():
-                name.terminate()
 
     def test_new_node_joining_copies_data(self):
         ports = [3342, 3343]
